requirements.txt
# ASTROPATH - Smart Road Damage Reporting & Rapid Response System
# Complete Dependencies for Full System

# Computer Vision & Deep Learning
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
tensorflow>=2.13.0
keras>=2.13.0
numpy>=1.23.0
scipy>=1.10.0

# Web Framework & API
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
werkzeug>=2.3.0

# Data Processing
scikit-learn>=1.3.0
pandas>=2.0.0
Pillow>=10.0.0

# Geolocation & GPS
geocoder>=1.38.1
gpsd-py3>=0.0.1
pynmea2>=1.20.0
pyserial>=3.5

# ESP32-CAM Streaming
urllib3>=1.26.0

# Database
# (sqlite3 is built-in with Python)

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Optional: Raspberry Pi / Edge Deployment
# tensorflow-lite  # TFLite runtime for Pi
# openvino>=2023.0  # For Intel VPU acceleration
# ncnn-python>=0.0.1  # For NCNN inference

# Development & Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Documentation (optional)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

